Properly get X screen application is running on in appScreen
QDesktopWidget does not provide the X screen number. No matter on which
X screen an application is started, QDesktopWidget returns 0.

Therefore resolve the X screen through the native interface's x11
screen resource.

The documentation is improved to make it more obvious that this is
about the X screen and not the RandR screen.

Change-Id: I0f78a728769cbdfcac8b67b4fd5de0934f498027
Reviewed-by: Richard J. Moore <rich@kde.org>
Reviewed-by: David Faure <david.faure@kdab.com>