Do not return null Display or Connection when there are no screens
This is a followup change for recent changes in QtGui which now handles a
situation where there are no outputs, and hence a nullptr QScreen comes
haunting us.

Thanks to Martin Gräßlin for pointing out where the problem is, and to Richard
Moore for explaining how to fix it.

Change-Id: I8f70c7a0b232073d4edb55b29eb19dae9fc09413
Reviewed-by: Richard J. Moore <rich@kde.org>
Reviewed-by: Shawn Rutledge <shawn.rutledge@digia.com>