Fix return type of appRootWindow for Qt 6
This is minor, but it's a bit silly to cast down to a 32 bit type in
return and then cast it up to 64 bit again on the same line.
The returned value is most likely used to other x11 calls which expect
32 bit.
I think this is so minor that it doesn't even need a change log entry
for Qt 6.

Change-Id: I3b54394c7934eb62026dfcf241e983c20b7e7d9c
Reviewed-by: David Faure <david.faure@kdab.com>