Update commercial license headers
Updated header.COMM to the files in tqtc-qtx11extras. Examples, tests
or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is
that some autotests use hard coded line numbers
and a change in the line count causes failures
in tests.

Task-number: QTQAINFRA-4170
Change-Id: I110fd6e0975c35ff40e92a04cd5b13e246f71252
Reviewed-by: Jani Heikkinen <jani.heikkinen@qt.io>